+++ Tetris/Core/Src/tablero.c
@@ -3,20 +3,20 @@
 
 
 void Tablero_Init(Tablero_t *me) {
-    for (int i = 0; i < TABLERO_ALTO; i++) {
-        for (int j = 0; j < TABLERO_ANCHO; j++) {
+    for (int8_t i = 0; i < TABLERO_ALTO; i++) {
+        for (int8_t j = 0; j < TABLERO_ANCHO; j++) {
             me->matriz[i][j] = 0;
         }
     }
 }
 
 bool Tablero_VerificarColision(Tablero_t *me, Pieza_t *pieza) {
-    for (int i = 0; i < 4; i++) {
-        for (int j = 0; j < 4; j++) {
+    for (int8_t i = 0; i < 4; i++) {
+        for (int8_t j = 0; j < 4; j++) {
             // Solo comprobamos si en la matriz 4x4 de la pieza hay un bloque
             if (Pieza_GetBloque(pieza, i, j)) {
-                int realX = pieza->x + j;
-                int realY = pieza->y + i;
+                int8_t realX = pieza->x + j;
+                int8_t realY = pieza->y + i;
 
                 // Límites laterales
                 if (realX < 0 || realX >= TABLERO_ANCHO) return true;
@@ -35,11 +35,11 @@ bool Tablero_VerificarColision(Tablero_t *me, Pieza_t *pieza) {
 }
 
 void Tablero_FijarPieza(Tablero_t *me, Pieza_t *pieza) {
-    for (int i = 0; i < 4; i++) {
-        for (int j = 0; j < 4; j++) {
+    for (int8_t i = 0; i < 4; i++) {
+        for (int8_t j = 0; j < 4; j++) {
             if (Pieza_GetBloque(pieza, i, j)) {
-                int realX = pieza->x + j;
-                int realY = pieza->y + i;
+                int8_t realX = pieza->x + j;
+                int8_t realY = pieza->y + i;
 
                 if (realY >= 0 && realY < TABLERO_ALTO && realX >= 0 && realX < TABLERO_ANCHO) {
                     me->matriz[realY][realX] = 1;
@@ -50,11 +50,11 @@ void Tablero_FijarPieza(Tablero_t *me, Pieza_t *pieza) {
 }
 
 int Tablero_EliminarLineasCompletas(Tablero_t *me) {
-    int lineasEliminadas = 0;
+    int8_t lineasEliminadas = 0;
 
-    for (int i = TABLERO_ALTO - 1; i >= 0; i--) {
+    for (int8_t i = TABLERO_ALTO - 1; i >= 0; i--) {
         bool filaLlena = true;
-        for (int j = 0; j < TABLERO_ANCHO; j++) {
+        for (int8_t j = 0; j < TABLERO_ANCHO; j++) {
             if (me->matriz[i][j] == 0) {
                 filaLlena = false;
                 break;
@@ -64,13 +64,13 @@ int Tablero_EliminarLineasCompletas(Tablero_t *me) {
         if (filaLlena) {
             lineasEliminadas++;
             // Se bajan las filas superiores manualmente
-            for (int k = i; k > 0; k--) {
-                for (int col = 0; col < TABLERO_ANCHO; col++) {
+            for (int8_t k = i; k > 0; k--) {
+                for (int8_t col = 0; col < TABLERO_ANCHO; col++) {
                     me->matriz[k][col] = me->matriz[k-1][col];
                 }
             }
             // Se limpia la fila de arriba del todo
-            for (int col = 0; col < TABLERO_ANCHO; col++) {
+            for (int8_t col = 0; col < TABLERO_ANCHO; col++) {
                 me->matriz[0][col] = 0;
             }
             i++; // Volvemos a revisar esta fila por si la que bajó también está llena
